+++ src/components/AssociationEditor/menu/SchemaItem.vue
@@ -2,10 +2,14 @@
 import {ref, watch} from "vue";
 import {GenSchemaView, GenTableCommonView} from "../../../api/__generated/model/static";
 import {api} from "../../../api";
-import {useTableEditorGraphStore} from "../../../store/tableEditorGraph";
+import {useAssociationEditorGraphStore} from "../../../store/AssociationEditorGraphStore.ts";
 import Details from "../../common/Details.vue";
+import {Delete, Search} from "@element-plus/icons-vue";
+import {ElMessageBox} from 'element-plus'
+import TableItem from "./TableItem.vue";
+import SchemaIcon from "../../icons/database/SchemaIcon.vue";
 
-const store = useTableEditorGraphStore()
+const store = useAssociationEditorGraphStore()
 
 interface SchemaItemProps {
 	schema: GenSchemaView
@@ -32,10 +36,20 @@ watch(() => props.schema, () => {
 }, {immediate: true})
 
 const deleteSchema = () => {
-	api.schemaService.delete({ids: [props.schema.id]}).then(res => {
-		if (res > 0) {
-			emits("delete", props.schema.id)
+	ElMessageBox.confirm(
+		`确定要删除 ${props.schema.name} 吗？`,
+		{
+			confirmButtonText: 'Yes',
+			cancelButtonText: 'No',
+			icon: Delete,
+			type: "error"
 		}
+	).then(() => {
+		api.schemaService.delete({ids: [props.schema.id]}).then(res => {
+			if (res > 0) {
+				emits("delete", props.schema.id)
+			}
+		})
 	})
 }
 
@@ -51,35 +65,37 @@ const query = () => {
 		tables.value = res
 	})
 }
-
 </script>
 
 <template>
 	<Details>
 		<template #title>
-			<div style="height: 2.5em; line-height: 2.5em;">
-				<span
-					@click.prevent.stop="store.importSchema([...tables.map(table => table.id)])">{{
-						schema.name
-					}}</span>
-				<el-button @click.prevent.stop="deleteSchema">删除</el-button>
+			<div style="height: 1.8em; line-height: 1.8em;">
+				<el-text>
+					<SchemaIcon></SchemaIcon>
+
+					<el-button @click="store.loadSchema([...tables.map(table => table.id)])" link>
+						{{ schema.name }}
+					</el-button>
+
+					<el-button @click="deleteSchema" title="删除" :icon="Delete"
+							   type="danger" link>
+					</el-button>
+				</el-text>
 			</div>
 		</template>
-		<table style="padding-left: 1em;">
-			<tr>
-				<td colspan="2">
-					<el-input v-model="keywords" @change="query">
-						<template #append>
-							<el-button @click="query">搜索</el-button>
-						</template>
-					</el-input>
-				</td>
-			</tr>
-			<tr v-for="table in tables" :class="table.type"
-				@click.prevent.stop="store.importTable(table.id)">
-				<td>{{ table.name }}</td>
-				<td>{{ table.comment }}</td>
-			</tr>
-		</table>
+
+		<div style="padding-left: 0.5em;">
+			<el-input v-model="keywords" @change="query" clearable>
+				<template #append>
+					<el-button @click="query" title="搜索" :icon="Search"></el-button>
+				</template>
+			</el-input>
+			<ul style="padding: 0 0 0.5em 0.5em;">
+				<li v-for="table in tables">
+					<TableItem :table="table"></TableItem>
+				</li>
+			</ul>
+		</div>
 	</Details>
 </template>
